Completed the basic requirements of Google Homepage but needs refinement
The homepage is completed and might be janky compartively to the
original homepage but it is something in generally quite happy with.
The main challenge was to allow the search icon to follow the text input of the searchbox as I've decided to make the homepage slightly more dynamic with adjustable viewport height as well as some hover qualities. Will continue to refine the site in the future as I learn more things but I think it is time to move on to the next lesson. Will also challenge the harder assignment of the result page in the foreseeable future.

[Ticket:2-19/4/2021]